日线行情解析成交额信息时应该用float32类型接收，否则数据会解析有问题。

+++ packet/response.go
@@ -81,7 +81,7 @@ type StockDayItem struct {
      High       uint32
      Low        uint32
      Close      uint32
-     Amount     uint32
+     Amount     float32
      Volume     uint32
      Unknown1   uint32
 }
